requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
databases[postgresql,sqlite]>=0.9.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.3
python-multipart>=0.0.6
aiosqlite>=0.20.0
httpx>=0.26.0
numpy>=1.26.0
Pillow>=10.0.0

# ML - Detection + Tracking + Pose (YOLO11 + BoT-SORT built-in)
ultralytics>=8.3.114

# ML - Face Recognition + ReID (SCRFD + ArcFace)
insightface>=0.7.3
onnxruntime>=1.17.0

# ML - Demographics (MiVOLO v2 age+gender)
transformers>=4.36.0

# Video
opencv-python>=4.9.0
yt-dlp>=2024.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
